+++ src/pdftl/cli/main.py
@@ -94,16 +94,19 @@ def _find_help_command(cli_args):
     It searches topics in a specific order: special, operator, then option.
     """
     from pdftl.cli.help import (
+        TAG_PREFIX,
         find_operator_topic_command,
         find_option_topic_command,
         find_special_topic_command,
     )
 
+    tag_queries = [arg for arg in cli_args if arg.startswith(TAG_PREFIX)]
     help_topics = [arg for arg in cli_args if arg not in HELP_FLAGS]
     first_topic = help_topics[0].lower() if help_topics else None
     help_args = [arg for arg in cli_args if arg in HELP_FLAGS]
     return (
-        find_special_topic_command(first_topic)
+        (tag_queries and tag_queries[0])
+        or find_special_topic_command(first_topic)
         or find_operator_topic_command(help_topics)
         or find_option_topic_command(help_topics)
         or (len(help_args) > 1 and find_special_topic_command(help_args[1]))